Revert "Merge pull request #1 from sweshelo/feature/turbo"
This reverts commit f2ad037ceeca09f08b7a25727b60bd5a5f820554, reversing
changes made to a7e97ec48968495bb161607c33ba4dc269b23a90.
