+++ .github/workflows/build.yml
@@ -19,10 +19,10 @@ jobs:
           cache: npm
 
       - name: Create .env file
-        run: echo -e "${{ secrets.DOTENV }}" > apps/web/.env
+        run: echo "${{ secrets.DOTENV }}" > .env
 
       - name: npm-install
         run: npm i
 
       - name: npm-build
-        run: npx turbo run build
+        run: npm run build
