+++ package.json
@@ -1,16 +1,65 @@
 {
+  "name": "enma-v2",
+  "version": "0.1.0",
   "private": true,
-  "workspaces": [
-    "apps/*",
-    "packages/*"
-  ],
-  "packageManager": "npm@10.8.2",
-  "lint-staged": {
-    "apps/web/**/*.{ts,tsx,js,jsx}": [
-      "npx --filter=enma-v2-web lint-staged"
-    ]
+  "scripts": {
+    "dev": "next dev --turbopack",
+    "build": "next build",
+    "build:verify": "rimraf .next-verify && cross-env NEXT_BUILD_DIR=.next-verify next build && rimraf .next-verify",
+    "start": "next start",
+    "lint": "next lint --fix",
+    "format": "prettier --write .",
+    "prepare": "husky",
+    "check:deps": "depcruise src"
+  },
+  "dependencies": {
+    "@supabase/supabase-js": "^2.46.1",
+    "@vercel/analytics": "^1.4.1",
+    "@vercel/speed-insights": "^1.1.0",
+    "cheerio": "^1.0.0",
+    "classnames": "^2.5.1",
+    "date-fns": "^4.1.0",
+    "highcharts": "^11.4.8",
+    "highcharts-react-official": "^3.2.1",
+    "next": "15.0.3",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "recharts": "^2.13.0-alpha.5",
+    "supabase": "^1.223.10"
   },
   "devDependencies": {
-    "turbo": "^2.3.3"
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "@typescript-eslint/eslint-plugin": "^8.14.0",
+    "@typescript-eslint/parser": "^8.14.0",
+    "cross-env": "^7.0.3",
+    "dependency-cruiser": "^16.6.0",
+    "eslint": "^8.57.1",
+    "eslint-config-next": "15.0.3",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "husky": "^9.1.6",
+    "lint-staged": "^15.2.10",
+    "postcss": "^8",
+    "prettier": "^3.3.3",
+    "rimraf": "^6.0.1",
+    "sass": "^1.81.0",
+    "sass-loader": "^16.0.3",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5"
+  },
+  "volta": {
+    "node": "20.18.0"
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx}": [
+      "eslint --fix",
+      "prettier --write"
+    ],
+    "*.{json,md}": [
+      "prettier --write"
+    ]
   }
 }
