+++ src/components/common/shiny.tsx
