+++ src/components/stage-icon/index.tsx
