+++ src/constants/chara.ts
