+++ src/service/supabase/client.ts
@@ -1,0 +1,8 @@
+import { createClient } from '@supabase/supabase-js'
+
+import { Database } from '@/types/database.types'
+
+export const supabase = createClient<Database>(
+  process.env.SUPABASE_URL!,
+  process.env.SUPABASE_API_KEY!
+)
