--- turbo.json
@@ -1,20 +1,0 @@
-{
-  "tasks": {
-    "build": {
-      "outputs": [
-        ".next/**"
-      ],
-      "env": [
-        "SUPABASE_PASSWORD",
-        "SUPABASE_URL",
-        "SUPABASE_API_KEY"
-      ]
-    },
-    "build:verify": {},
-    "lint": {},
-    "test": {},
-    "dev": {
-      "cache": false
-    }
-  }
-}
